+++ lib/size_providers.dart
@@ -1,13 +1,12 @@
 import 'package:flutter/widgets.dart';
 
-typedef Size CardSizeProvider(BuilderData data);
-typedef Size SeparatorSizeProvider(BuilderData data);
+typedef Size CardSizeProvider(int position, BuilderData data);
+typedef Size SeparatorSizeProvider(int position, BuilderData data);
 
 class BuilderData {
-  final int current;
   final int center;
   final int next;
   final double progress;
 
-  BuilderData(this.current, this.center, this.next, this.progress);
+  BuilderData(this.center, this.next, this.progress);
 }