+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Changelog
+
+## [0.1.1] - 24.10.2018
+
+* Explicitly provides current item position, does not hide it in builder data
+* Removes unnessesary `setState` calls from snip animation
+
 ## [0.1.0] - 21.10.2018
 
 * It's alive!